third_party/blink/web_tests/external/wpt/css/css-inline/text-box-trim/parsing/text-box-edge-computed.html

<!DOCTYPE html>
<title>text-box getComputedStyle() returns correctly resolved values</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#text-edges">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<body>
<div id="target"></div>
<script>
test_computed_value('text-box-edge', 'auto');
test_computed_value('text-box-edge', 'text');
test_computed_value('text-box-edge', 'ideographic');
test_computed_value('text-box-edge', 'ideographic-ink');
test_computed_value('text-box-edge', 'ex');
test_computed_value('text-box-edge', 'cap');
test_computed_value('text-box-edge', 'text alphabetic');
test_computed_value('text-box-edge', 'text text', 'text');
test_computed_value('text-box-edge', 'cap alphabetic');
test_computed_value('text-box-edge', 'cap text', 'cap');
test_computed_value('text-box-edge', 'ex alphabetic');
test_computed_value('text-box-edge', 'ex text', 'ex');
test_computed_value('text-box-edge', 'ex ideographic');
test_computed_value('text-box-edge', 'ideographic-ink alphabetic');

</script>
</body>
