third_party/blink/web_tests/external/wpt/css/css-inline/text-box-trim/parsing/text-box-edge-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-box-edge value 'ideographic'
  assert_true: 'ideographic' is a supported value for text-box-edge. expected true got false
[FAIL] Property text-box-edge value 'ideographic-ink'
  assert_true: 'ideographic-ink' is a supported value for text-box-edge. expected true got false
[FAIL] Property text-box-edge value 'ex ideographic'
  assert_true: 'ex ideographic' is a supported value for text-box-edge. expected true got false
[FAIL] Property text-box-edge value 'ideographic-ink alphabetic'
  assert_true: 'ideographic-ink alphabetic' is a supported value for text-box-edge. expected true got false
Harness: the test ran to completion.
